+++ packages/node-core/src/indexer/fetch.service.ts
@@ -36,8 +36,8 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
 
   // The rough interval at which new blocks are produced
   protected abstract getChainInterval(): Promise<number>;
-  // This return modulos block (that provided in the handler filter) within a given block range
-  protected abstract getModulos(startHeight: number, endHeight: number): number[];
+  // This return all modulo numbers in the filters of all dataSources
+  protected abstract getAllModuloNumbers(): number[];
 
   protected abstract initBlockDispatcher(): Promise<void>;
 
@@ -180,13 +180,17 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
     await this.fillNextBlockBuffer(initBlockHeight);
   }
 
-  // get all modulo numbers with a specific block ranges
-  getModuloBlocks(startHeight: number, endHeight: number): number[] {
-    // valid modulo numbers from handler filters
-    const modulos = this.getModulos(startHeight, endHeight);
+  // This return modulos numbers (those provided in the handler filters) within a given block range
+  private getModuloNumbers(startHeight: number, endHeight: number): number[] {
+    const allModulos = this.getAllModuloNumbers();
+    return allModulos.filter((n) => n >= startHeight && n <= endHeight);
+  }
+
+  // get all modulo numbers with a specific block ranges, and given moduloNumbers (from filters)
+  private getModuloBlocks(startHeight: number, endHeight: number, moduloNumbers: number[]): number[] {
     const moduloBlocks: number[] = [];
-    for (let i = startHeight; i < endHeight; i++) {
-      if (modulos.find((m) => i % m === 0)) {
+    for (let i = startHeight; i <= endHeight; i++) {
+      if (moduloNumbers.find((m) => i % m === 0)) {
         moduloBlocks.push(i);
       }
     }
@@ -197,14 +201,17 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
    *
    * @param startBlockHeight
    * @param endBlockHeight is either FinalizedHeight or BestHeight, ensure ModuloBlocks not greater than this number
+   * @param moduloNumbers valid modulo number from handler filters
    */
-  getEnqueuedModuloBlocks(startBlockHeight: number, endBlockHeight: number): (IBlock<FB> | number)[] {
+  private getEnqueuedModuloBlocks(
+    startBlockHeight: number,
+    endBlockHeight: number,
+    moduloNumbers: number[]
+  ): (IBlock<FB> | number)[] {
     return this.getModuloBlocks(
       startBlockHeight,
-      Math.min(
-        this.nodeConfig.batchSize * Math.max(...this.getModulos(startBlockHeight, endBlockHeight)) + startBlockHeight,
-        endBlockHeight
-      )
+      Math.min(this.nodeConfig.batchSize * Math.max(...moduloNumbers) + startBlockHeight, endBlockHeight),
+      moduloNumbers
     ).slice(0, this.nodeConfig.batchSize);
   }
 
@@ -240,6 +247,23 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
         continue;
       }
 
+      // Find current datasources to check if they are only modulo filters, this also limits the range to the current DS set.
+      const details = this.projectService.getDataSourcesMap().getDetails(startBlockHeight);
+      assert(details, `Datasources not found for height ${startBlockHeight}`);
+      const {endHeight: rangeEndHeight, value: relevantDS} = details;
+      // Min of current ds endHeight
+      const minDsEndHeight = Math.min(...relevantDS.map((d) => d?.endBlock ?? Number.MAX_SAFE_INTEGER));
+      // Modulo end height should be min of current ds endHeight,
+      // if not defined, use rangeEndHeight (next ds startHeight -1), otherwise use latestHeight
+      // TODO, this is better handle from within blockHeight maps, use rangeEndHeight
+      const moduloEndHeight = Math.min(minDsEndHeight, rangeEndHeight ?? Number.MAX_SAFE_INTEGER, latestHeight);
+      // Modulo start height should be
+      // Get modulo numbers from range within min of ds startHeight(if not defined should use startBlockHeight), to end of expected moduloEndHeight
+      const moduloNumbers = this.getModuloNumbers(
+        Math.min(...relevantDS.map((d) => d?.startBlock ?? startBlockHeight)),
+        moduloEndHeight
+      );
+
       if (
         this.useDictionary &&
         // TODO, do we still need to check useDictionary method here, this will
@@ -260,7 +284,7 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
           if (dictionary) {
             const {batchBlocks} = dictionary;
             // the last block returned from batch should have max height in this batch
-            const moduloBlocks = this.getModuloBlocks(startBlockHeight, dictionary.lastBufferedHeight);
+            const moduloBlocks = this.getModuloBlocks(startBlockHeight, dictionary.lastBufferedHeight, moduloNumbers);
             const mergedBlocks = mergeNumAndBlocks(moduloBlocks, batchBlocks);
             if (mergedBlocks.length === 0) {
               // There we're no blocks in this query range, we can set a new height we're up to
@@ -279,25 +303,16 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
         }
       } else {
         const endHeight = this.nextEndBlockHeight(startBlockHeight, scaledBatchSize);
-        // Find current datasources to check if they are only modulo filters, this also limits the range to the current DS set.
-        const details = this.projectService.getDataSourcesMap().getDetails(startBlockHeight);
-        assert(details, `Datasources not found for height ${startBlockHeight}`);
-        const {endHeight: rangeEndHeight, value: relevantDS} = details;
+
         const handlers = [...relevantDS.map((ds) => ds.mapping.handlers)].flat();
-        // Min of current ds endHeight
-        const minDsEndHeight = Math.min(...relevantDS.map((d) => d?.endBlock ?? Number.MAX_SAFE_INTEGER));
-        // Modulo end height should be min of current ds endHeight,
-        // if not defined, use rangeEndHeight (next ds startHeight -1), otherwise use latestHeight
-        // TODO, this is better handle from within blockHeight maps, use rangeEndHeight
-        const moduloEndHeight = Math.min(minDsEndHeight, rangeEndHeight ?? Number.MAX_SAFE_INTEGER, latestHeight);
 
         const enqueuingBlocks =
           // This check whether within blockRange has only block modulos handler or with others
-          // If modulos hanlder only, we will enqueue only modulos number,
-          // If has other handlers, we will enqueue every block
-          handlers.length && this.getModulos(startBlockHeight, moduloEndHeight).length === handlers.length
-            ? this.getEnqueuedModuloBlocks(startBlockHeight, moduloEndHeight)
-            : // Should also min include dataSource endHeight and endHeight (calculated by batch size)
+          // If are modulos hanlders only, then number of moduloNumbers should be match number of with handlers, we will enqueue only modulos,
+          // If have other handlers, we will enqueue every block
+          handlers.length && moduloNumbers.length === handlers.length
+            ? this.getEnqueuedModuloBlocks(startBlockHeight, moduloEndHeight, moduloNumbers)
+            : // EndHeight should also consider dataSource endHeight and estimated endHeight (calculated by batch size)
               range(startBlockHeight, Math.min(rangeEndHeight ?? Number.MAX_SAFE_INTEGER, endHeight) + 1);
 
         await this.enqueueBlocks(enqueuingBlocks, latestHeight);